+++ src/java/com/wolfssl/provider/jsse/WolfSSLSocket.java
@@ -73,7 +73,12 @@ public class WolfSSLSocket extends SSLSocket {
 
     private ArrayList<HandshakeCompletedListener> hsListeners = null;
     private WolfSSLDebug debug;
+
     protected volatile boolean handshakeInitCalled = false;
+    protected volatile boolean handshakeComplete = false;
+
+    /* lock for handshakInitCalled and handshakeComplete */
+    final private Object handshakeLock = new Object();
 
     public WolfSSLSocket(com.wolfssl.WolfSSLContext context,
            WolfSSLAuthStore authStore, WolfSSLParameters params,
@@ -512,12 +517,19 @@ public class WolfSSLSocket extends SSLSocket {
     synchronized public void startHandshake() throws IOException {
         int ret;
 
-        /* will throw SSLHandshakeException if session creation is
-           not allowed */
-        EngineHelper.initHandshake();
-        handshakeInitCalled = true;
+        synchronized (handshakeLock) {
+            if (handshakeInitCalled == true || handshakeComplete == true) {
+                /* handshake already started or finished */
+                return;
+            }
 
-        ret = EngineHelper.doHandshake();
+            /* will throw SSLHandshakeException if session creation is
+               not allowed */
+            EngineHelper.initHandshake();
+            handshakeInitCalled = true;
+
+            ret = EngineHelper.doHandshake();
+        }
 
         if (ret != WolfSSL.SSL_SUCCESS) {
 
@@ -526,6 +538,11 @@ public class WolfSSLSocket extends SSLSocket {
 
             throw new SSLHandshakeException(errStr + " (error code: " +
                 err + ")");
+        }
+
+        synchronized (handshakeLock) {
+            /* mark handshake completed */
+            handshakeComplete = true;
         }
 
         /* notify handshake completed listeners */
@@ -891,5 +908,170 @@ public class WolfSSLSocket extends SSLSocket {
             return ret;
         }
     }
+
+    class WolfSSLInputStream extends InputStream {
+
+        private WolfSSLSession ssl;
+        private WolfSSLSocket  socket;
+        final private Object readLock = new Object();
+
+        public WolfSSLInputStream(WolfSSLSession ssl, WolfSSLSocket socket) {
+            this.ssl = ssl;
+            this.socket = socket; /* parent socket */
+        }
+
+        @Override
+        public int read() throws IOException {
+
+            int ret = 0;
+            byte[] data = new byte[1];
+
+            try {
+                ret = this.read(data, 0, 1);
+
+            } catch (NullPointerException ne) {
+                throw new IOException(ne);
+
+            } catch (IndexOutOfBoundsException ioe) {
+                throw new IndexOutOfBoundsException(ioe.toString());
+            }
+
+            return (data[0] & 0xFF);
+        }
+
+        public int read(byte[] b) throws NullPointerException, IOException {
+
+            return this.read(b, 0, b.length);
+        }
+
+        public int read(byte[] b, int off, int len)
+            throws NullPointerException, IndexOutOfBoundsException, IOException {
+
+            int ret = 0;
+            byte[] data = null;
+
+            if (b == null) {
+                throw new NullPointerException("Input array is null");
+            }
+
+            /* do handshake if not completed yet, handles synchronization */
+            if (WolfSSLSocket.this.handshakeComplete == false) {
+                socket.startHandshake();
+            }
+
+
+            if (b.length == 0 || len == 0) {
+                return 0;
+            }
+
+            if (off < 0 || len < 0 || len > (b.length - off)) {
+                throw new IndexOutOfBoundsException("Array index out of bounds");
+            }
+
+            if (off != 0) {
+                /* create new tmp buffer to read data into */
+                data = new byte[len];
+            } else {
+                data = b;
+            }
+
+            synchronized (readLock) {
+                try {
+                    ret = ssl.read(data, len);
+                    if (ret <= 0) {
+                        int err = ssl.getError(ret);
+                        String errStr = WolfSSL.getErrorString(err);
+
+                        /* received CloseNotify, InputStream should return "-1"
+                           when there is no more data */
+                        if (err == WolfSSL.SSL_ERROR_ZERO_RETURN) {
+                            return -1;
+                        }
+
+                        throw new IOException("Native wolfSSL read() failed: " +
+                            errStr + " (error code: " + err + ")");
+                    }
+
+                } catch (IllegalStateException e) {
+                    throw new IOException(e);
+                }
+
+                if (off != 0) {
+                    /* copy data into original array at offset */
+                    System.arraycopy(data, 0, b, off, ret);
+                }
+
+                /* return number of bytes read */
+                return ret;
+            }
+        }
+    } /* end WolfSSLInputStream inner class */
+
+    class WolfSSLOutputStream extends OutputStream {
+
+        private WolfSSLSession ssl;
+        private WolfSSLSocket  socket;
+        final private Object writeLock = new Object();
+
+        public WolfSSLOutputStream(WolfSSLSession ssl, WolfSSLSocket socket) {
+            this.ssl = ssl;
+            this.socket = socket; /* parent socket */
+        }
+
+        public void write(int b) throws IOException {
+            byte[] data = new byte[1];
+            data[0] = (byte)(b & 0xFF);
+
+            this.write(data, 0, 1);
+        }
+
+        public void write(byte[] b) throws IOException {
+            this.write(b, 0, b.length);
+        }
+
+        public void write(byte[] b, int off, int len) throws IOException {
+
+            int ret;
+            byte[] data = null;
+
+            if (b == null) {
+                throw new NullPointerException("Input array is null");
+            }
+
+            /* do handshake if not completed yet, handles synchronization */
+            if (WolfSSLSocket.this.handshakeComplete == false) {
+                socket.startHandshake();
+            }
+
+            if (off < 0 || len < 0 || (off + len) > b.length) {
+                throw new IndexOutOfBoundsException("Array index out of bounds");
+            }
+
+            if (off != 0) {
+                data = new byte[len];
+                System.arraycopy(b, off, data, 0, len);
+            } else {
+                data = b;
+            }
+
+            synchronized (writeLock) {
+                try {
+                    ret = ssl.write(data, len);
+
+                    if (ret <= 0) {
+                        int err = ssl.getError(ret);
+                        String errStr = WolfSSL.getErrorString(err);
+
+                        throw new IOException("Native wolfSSL write() failed: " +
+                            errStr + " (error code: " + err + ")");
+                    }
+
+                } catch (IllegalStateException e) {
+                    throw new IOException(e);
+                }
+            }
+        }
+    } /* end WolfSSLOutputStream inner class */
+
 }
